+++ build.sh
@@ -18,6 +18,8 @@
 SCRIPT_DIR=$(cd -P "$(dirname "$0")" >/dev/null 2>&1 && pwd)
 SCRIPT_FILE="${SCRIPT_DIR}/$(basename "$0")"
 
+mvn_cmd=$SCRIPT_DIR/mvnw
+
 ##
 ## Local Development
 ##
@@ -26,12 +28,12 @@ SCRIPT_FILE="${SCRIPT_DIR}/$(basename "$0")"
 
 ## clean - Cleans the build output
 function clean() {
-  mvn clean
+  $mvn_cmd clean
 }
 
 ## build - Builds the project without running tests.
 function build() {
-   mvn install -DskipTests=true
+   $mvn_cmd install -DskipTests=true
 }
 
 ## test - Runs local unit tests.
@@ -39,10 +41,14 @@ function test() {
   if [[ "$(uname -s)" == "Darwin" ]]; then
     echo "macOS detected. Setting up IP aliases for tests."
     echo "You may be prompted for your password to run sudo."
-    sudo ifconfig lo0 alias 127.0.0.2 up
-    sudo ifconfig lo0 alias 127.0.0.3 up
+    if ! ifconfig lo0 | grep -q 127.0.0.2 ; then
+      sudo ifconfig lo0 alias 127.0.0.2 up
+    fi
+    if ! ifconfig lo0 | grep -q 127.0.0.3 ; then
+      sudo ifconfig lo0 alias 127.0.0.3 up
+    fi
   fi
-  mvn -P coverage test
+  $mvn_cmd -P coverage test
 }
 
 ## e2e - Runs end-to-end integration tests.
@@ -66,18 +72,18 @@ function e2e_graalvm() {
 
 ## fix - Fixes java code format.
 function fix() {
-  mvn com.coveo:fmt-maven-plugin:format
+  $mvn_cmd com.coveo:fmt-maven-plugin:format
 }
 
 ## lint - runs the java lint
 function lint() {
-  mvn -P lint install -DskipTests=true
+  $mvn_cmd -P lint install -DskipTests=true
 }
 
 
 ## deps - updates dependencies to the latest version
 function deps() {
-  mvn versions:use-latest-versions
+  $mvn_cmd versions:use-latest-versions
   find . -name 'pom.xml.versionsBackup' -print0 | xargs -0 rm -f
 }
 
@@ -91,7 +97,7 @@ function write_e2e_env(){
   secret_vars=(
     MYSQL_CONNECTION_NAME=MYSQL_CONNECTION_NAME
     MYSQL_USER=MYSQL_USER
-    MYSQL_USER_IAM=MYSQL_USER_IAM_GO
+    IMPERSONATED_USER=IMPERSONATED_USER
     MYSQL_PASS=MYSQL_PASS
     MYSQL_DB=MYSQL_DB
     MYSQL_MCP_CONNECTION_NAME=MYSQL_MCP_CONNECTION_NAME
@@ -105,8 +111,8 @@ function write_e2e_env(){
     POSTGRES_CAS_PASS=POSTGRES_CAS_PASS
     POSTGRES_CUSTOMER_CAS_CONNECTION_NAME=POSTGRES_CUSTOMER_CAS_CONNECTION_NAME
     POSTGRES_CUSTOMER_CAS_PASS=POSTGRES_CUSTOMER_CAS_PASS
-    POSTGRES_CUSTOMER_CAS_DOMAIN_NAME=POSTGRES_CUSTOMER_CAS_DOMAIN_NAME
-    POSTGRES_CUSTOMER_CAS_INVALID_DOMAIN_NAME=POSTGRES_CUSTOMER_CAS_INVALID_DOMAIN_NAME
+    POSTGRES_CUSTOMER_CAS_PASS_VALID_DOMAIN_NAME=POSTGRES_CUSTOMER_CAS_DOMAIN_NAME
+    POSTGRES_CUSTOMER_CAS_PASS_INVALID_DOMAIN_NAME=POSTGRES_CUSTOMER_CAS_INVALID_DOMAIN_NAME
     POSTGRES_MCP_CONNECTION_NAME=POSTGRES_MCP_CONNECTION_NAME
     POSTGRES_MCP_PASS=POSTGRES_MCP_PASS
     SQLSERVER_CONNECTION_NAME=SQLSERVER_CONNECTION_NAME
@@ -131,6 +137,10 @@ function write_e2e_env(){
     val=$(gcloud secrets versions access latest --project "$TEST_PROJECT" --secret="$secret_name")
     echo "export $env_var_name='$val'"
   done
+
+  echo "export MYSQL_IAM_USER='$(whoami)'"
+  echo "export POSTGRES_IAM_USER='$(whoami)@google.com'"
+
   } > "$outfile"
 
 }