+++ r2dbc/core/src/main/java/com/google/cloud/sql/core/CloudSqlConnectionFactory.java
@@ -27,7 +27,6 @@ import io.r2dbc.spi.ConnectionFactoryProvider;
 import java.util.function.Supplier;
 import org.reactivestreams.Publisher;
 import reactor.core.publisher.Mono;
-import reactor.util.annotation.NonNull;
 
 /** {@link ConnectionFactory} for accessing Cloud SQL instances via R2DBC protocol. */
 public class CloudSqlConnectionFactory implements ConnectionFactory {
@@ -48,20 +47,17 @@ public class CloudSqlConnectionFactory implements ConnectionFactory {
   }
 
   @Override
-  @NonNull
   public Publisher<? extends Connection> create() {
     String hostIp =
         InternalConnectorRegistry.getInstance()
             .getConnectionMetadata(config)
             .getPreferredIpAddress();
     builder.option(HOST, hostIp).option(PORT, SERVER_PROXY_PORT);
-
     return Mono.from(supplier.get().create(builder.build()).create())
         .map(c -> new CloudSqlConnection(config, c));
   }
 
   @Override
-  @NonNull
   public ConnectionFactoryMetadata getMetadata() {
     String hostIp =
         InternalConnectorRegistry.getInstance()